CN-4874 Refactor
