+++ app/components/screens/HODLViewCode/HODLViewCode.styles.js
